Fixed some bugs in the new double-math volume department...

+++ projects/drummaster/rev2/src/pad/Cymbal.cpp
@@ -13,7 +13,7 @@ void Cymbal::poll(){
 		}
 	}
 	else {
-		uint16_t volume = readPiezo(piezoMuxIndex);
+		double volume = readPiezo(piezoMuxIndex);
 		if (volume){
 			play(volume);
 		}
